Undo unification of getting variable info for ease of testing and general simplification
