+++ packages/grafana-data/src/field/correlationUtils.ts
@@ -1,9 +1,7 @@
-import { uniqBy } from 'lodash';
 import logfmt from 'logfmt';
 
 import { ScopedVars } from '../types/ScopedVars';
-import { DataLinkTransformationConfig, SupportedTransformationType, VariableInterpolation } from '../types/dataLink';
-import { InterpolateFunction } from '../types/panel';
+import { DataLinkTransformationConfig, SupportedTransformationType } from '../types/dataLink';
 
 export const safeStringifyValue = (value: unknown, space?: number) => {
   if (value === undefined || value === null) {
@@ -52,7 +50,7 @@ export const getTransformationVars = (
 };
 
 // See https://grafana.com/docs/grafana/latest/dashboards/variables/add-template-variables/#global-variables
-export const builtInVariablesGlobal = [
+export const builtInVariables = [
   '__from',
   '__to',
   '__interval',
@@ -63,48 +61,7 @@ export const builtInVariablesGlobal = [
   '__rate_interval',
   '__timeFilter',
   'timeFilter',
+  // These are only applicable in dashboards so should not affect this for Explore
+  // '__dashboard',
+  //'__name',
 ];
-
-// These are only applicable in dashboards so should not affect Explore
-const builtInVariablesDashboards = ['__dashboard', '__name'];
-
-export const builtInVariables = [...builtInVariablesGlobal, ...builtInVariablesDashboards];
-
-/**
- * Use variable map from templateSrv to determine if all variables have values
- * @param query
- * @param scopedVars
- */
-export function getVariableUsageInfo(
-  query: object,
-  scopedVars: ScopedVars,
-  replaceFn: InterpolateFunction
-): { variables: VariableInterpolation[]; allVariablesDefined: boolean } {
-  let variables: VariableInterpolation[] = [];
-  // This adds info to the variables array while interpolating
-  replaceFn(getStringsFromObject(query), scopedVars, undefined, variables);
-  variables = uniqBy(variables, 'variableName');
-  return {
-    variables: variables,
-    allVariablesDefined: variables
-      // We filter out builtin variables as they should be always defined but sometimes only later, like
-      // __range_interval which is defined in prometheus at query time.
-      .filter((v) => !builtInVariables.includes(v.variableName))
-      .every((variable) => variable.found),
-  };
-}
-
-// Recursively get all strings from an object into a simple list with space as separator.
-function getStringsFromObject(obj: Object): string {
-  let acc = '';
-  let k: keyof typeof obj;
-
-  for (k in obj) {
-    if (typeof obj[k] === 'string') {
-      acc += ' ' + obj[k];
-    } else if (typeof obj[k] === 'object') {
-      acc += ' ' + getStringsFromObject(obj[k]);
-    }
-  }
-  return acc;
-}
