+++ public/app/features/explore/utils/links.ts
@@ -1,4 +1,4 @@
-import { first } from 'lodash';
+import { first, uniqBy } from 'lodash';
 import { useCallback } from 'react';
 
 import {
@@ -19,9 +19,9 @@ import {
   DataLinkPostProcessor,
   ExploreUrlState,
   urlUtil,
-  VariableInterpolation,
   getTransformationVars,
-  getVariableUsageInfo,
+  VariableInterpolation,
+  builtInVariables,
 } from '@grafana/data';
 import { getTemplateSrv, reportInteraction } from '@grafana/runtime';
 import { DataQuery } from '@grafana/schema';
@@ -42,6 +42,17 @@ const dataLinkHasRequiredPermissionsFilter = (link: DataLink) => {
  * be passed back to the visualization.
  */
 const DATA_LINK_FILTERS: DataLinkFilter[] = [dataLinkHasRequiredPermissionsFilter];
+
+/**
+ * This extension of the LinkModel was done to support correlations, which need the variables' names
+ * and values split out for display purposes
+ *
+ * Correlations are internal links only so the variables property will always be defined (but possibly empty)
+ * for internal links and undefined for non-internal links
+ */
+export interface ExploreFieldLinkModel extends LinkModel<Field> {
+  variables: VariableInterpolation[];
+}
 
 const DATA_LINK_USAGE_KEY = 'grafana_data_link_clicked';
 
@@ -100,7 +111,7 @@ export const getFieldLinksForExplore = (options: {
   dataFrame?: DataFrame;
   // if not provided, field.config.links are used
   linksToProcess?: DataLink[];
-}): LinkModel[] => {
+}): ExploreFieldLinkModel[] => {
   const { field, vars, splitOpenFn, range, rowIndex, dataFrame } = options;
   const scopedVars: ScopedVars = { ...(vars || {}) };
   scopedVars['__value'] = {
@@ -149,7 +160,7 @@ export const getFieldLinksForExplore = (options: {
       return DATA_LINK_FILTERS.every((filter) => filter(link, scopedVars));
     });
 
-    const fieldLinks: Array<LinkModel | undefined> = links.map((link) => {
+    const fieldLinks = links.map((link) => {
       let internalLinkSpecificVars: ScopedVars = {};
       if (link.meta?.transformations) {
         link.meta?.transformations.forEach((transformation) => {
@@ -168,9 +179,8 @@ export const getFieldLinksForExplore = (options: {
         });
       }
 
-      const replaceFn = getTemplateSrv().replace.bind(getTemplateSrv());
       const allVars = { ...scopedVars, ...internalLinkSpecificVars };
-      const variableData = getVariableUsageInfo(link, allVars, replaceFn);
+      const variableData = getVariableUsageInfo(link, allVars);
       let variables: VariableInterpolation[] = [];
 
       // if the link has no variables (static link), add it with the right key but an empty value so we know what field the static link is associated with
@@ -183,7 +193,7 @@ export const getFieldLinksForExplore = (options: {
       if (variableData.allVariablesDefined) {
         if (!link.internal) {
           const replace: InterpolateFunction = (value, vars) =>
-            replaceFn(value, { ...vars, ...allVars, ...scopedVars });
+            getTemplateSrv().replace(value, { ...vars, ...allVars, ...scopedVars });
 
           const linkModel = getLinkSrv().getDataLinkUIModel(link, replace, field);
           if (!linkModel.title) {
@@ -218,7 +228,7 @@ export const getFieldLinksForExplore = (options: {
         return undefined;
       }
     });
-    return fieldLinks.filter((link): link is LinkModel => !!link);
+    return fieldLinks.filter((link): link is ExploreFieldLinkModel => !!link);
   }
   return [];
 };
@@ -267,6 +277,47 @@ export function useLinks(range: TimeRange, splitOpenFn?: SplitOpen) {
     },
     [range, splitOpenFn]
   );
+}
+
+/**
+ * Use variable map from templateSrv to determine if all variables have values
+ *
+ * Note: There is a similar function in dashboard/state/PanelModel but this does not actually do the replacement. Maybe merge at some point.
+ * @param query
+ * @param scopedVars
+ */
+export function getVariableUsageInfo(
+  query: object,
+  scopedVars: ScopedVars
+): { variables: VariableInterpolation[]; allVariablesDefined: boolean } {
+  let variables: VariableInterpolation[] = [];
+  const replaceFn = getTemplateSrv().replace.bind(getTemplateSrv());
+  // This adds info to the variables array while interpolating
+  replaceFn(getStringsFromObject(query), scopedVars, undefined, variables);
+  variables = uniqBy(variables, 'variableName');
+  return {
+    variables: variables,
+    allVariablesDefined: variables
+      // We filter out builtin variables as they should be always defined but sometimes only later, like
+      // __range_interval which is defined in prometheus at query time.
+      .filter((v) => !builtInVariables.includes(v.variableName))
+      .every((variable) => variable.found),
+  };
+}
+
+// Recursively get all strings from an object into a simple list with space as separator.
+function getStringsFromObject(obj: Object): string {
+  let acc = '';
+  let k: keyof typeof obj;
+
+  for (k in obj) {
+    if (typeof obj[k] === 'string') {
+      acc += ' ' + obj[k];
+    } else if (typeof obj[k] === 'object') {
+      acc += ' ' + getStringsFromObject(obj[k]);
+    }
+  }
+  return acc;
 }
 
 type StateEntry = [string, ExploreItemState];
